echo: fold storage-metrics/GC helpers into EchoHost methods
Address review: replace the free functions in space-storage.ts and
garbage-collector.ts with private methods on EchoHost (the db host) —
_collectSpaceDocuments / _allSpaceDocumentIds / _countSpaceObjects /
_unlinkDeletedObjects / _wipeUnreachableDocuments. Behaviour is unchanged;
this just moves the logic onto the host that already owns getSpaceStats and
runGarbageCollection. Deletes the two helper files.
